+++ ChangeLog.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+- added useTypeHint option to VPack.Builder
+
 ## [2.3.1] - 2020-05-05
 
 - shaded jackson dependency